TestRangeExpr
